+++ pkg/front_end/lib/src/fasta/kernel/body_builder.dart
@@ -2616,19 +2616,32 @@ class BodyBuilder extends StackListenerImpl
     reportIfNotEnabled(
         libraryFeatures.patterns, token.charOffset, token.charCount);
     // ignore: unused_local_variable
-    Pattern left = toPattern(pop());
-    // ignore: unused_local_variable
     Pattern right = toPattern(pop());
+    // ignore: unused_local_variable
+    Pattern left = toPattern(pop());
     // TODO(johnniwinther): Create a binary pattern.
 
     String operator = token.lexeme;
-    BinaryPatternKind kind;
     switch (operator) {
       case '&':
-        kind = BinaryPatternKind.and;
+        push(new AndPattern(left, right, token.charOffset));
         break;
       case '|':
-        kind = BinaryPatternKind.or;
+        Map<String, VariableDeclaration> leftVariablesByName = {
+          for (VariableDeclaration leftVariable in left.declaredVariables)
+            leftVariable.name!: leftVariable
+        };
+        if (!leftVariablesByName.keys.toSet().containsAll(
+                right.declaredVariables.map((variable) => variable.name!)) &&
+            leftVariablesByName.length == right.declaredVariables.length) {
+          // TODO(cstefantsova): Report a compile-time error.
+        }
+        push(new OrPattern(left, right, token.charOffset,
+            orPatternJointVariables: [
+              for (VariableDeclaration leftVariable in left.declaredVariables)
+                forest.createVariableDeclaration(
+                    leftVariable.fileOffset, leftVariable.name)
+            ]));
         break;
       default:
         internalProblem(
@@ -2637,7 +2650,6 @@ class BodyBuilder extends StackListenerImpl
             token.charOffset,
             uri);
     }
-    push(new BinaryPattern(left, kind, right, token.charOffset));
   }
 
   void doBinaryExpression(Token token) {
@@ -4408,8 +4420,8 @@ class BodyBuilder extends StackListenerImpl
         ValueKinds.Pattern,
       ])
     ]));
-    Pattern key = toPattern(pop());
     Pattern value = toPattern(pop());
+    Pattern key = toPattern(pop());
     push(new MapPatternEntry(key, value, colon.charOffset));
   }
 
@@ -7615,16 +7627,7 @@ class BodyBuilder extends StackListenerImpl
       Message message, String className,
       [int charOffset = -1]) {
     addProblemErrorIfConst(message, charOffset, className.length);
-    // TODO(ahe): The following doesn't make sense to Analyzer AST.
-    MemberBuilder constructor =
-        libraryBuilder.loader.getAbstractClassInstantiationError();
-    Expression invocation = buildStaticInvocation(
-        constructor.member,
-        forest.createArguments(charOffset,
-            <Expression>[forest.createStringLiteral(charOffset, className)]),
-        constness: Constness.explicitNew,
-        charOffset: charOffset);
-    return forest.createThrow(charOffset, invocation);
+    return new InvalidExpression(message.problemMessage);
   }
 
   Statement buildProblemStatement(Message message, int charOffset,