+++ pkg/front_end/lib/src/fasta/kernel/kernel_target.dart
@@ -153,7 +153,6 @@ class KernelTarget extends TargetImplementation {
   /// Shared with [CompilerContext].
   final Map<Uri, Source> uriToSource = CompilerContext.current.uriToSource;
 
-  MemberBuilder? _cachedAbstractClassInstantiationError;
   MemberBuilder? _cachedCompileTimeError;
   MemberBuilder? _cachedDuplicatedFieldInitializerError;
   MemberBuilder? _cachedNativeAnnotation;
@@ -186,17 +185,6 @@ class KernelTarget extends TargetImplementation {
   }
 
   Uri? translateUri(Uri uri) => uriTranslator.translate(uri);
-
-  /// Returns a reference to the constructor of
-  /// `AbstractClassInstantiationError` error.  The constructor is expected to
-  /// accept a single argument of type String, which is the name of the
-  /// abstract class.
-  // TODO: Use some other error before `AbstractClassInstantiationError`
-  // is removed.
-  MemberBuilder getAbstractClassInstantiationError(Loader loader) {
-    return _cachedAbstractClassInstantiationError ??=
-        loader.coreLibrary.getConstructor("AbstractClassInstantiationError");
-  }
 
   /// Returns a reference to the constructor used for creating a compile-time
   /// error. The constructor is expected to accept a single argument of type